backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
google-generativeai
python-docx
pdfkit
python-multipart
